Not a number (NaN) handling in every brick action:
* Formula interpretation to NaN will be handled + test.
* Null-Formula tests.
* Stringfunctionhandling of Not-A-Number parameters.

+++ catroidTest/src/org/catrobat/catroid/test/formulaeditor/ParserTestObject.java
@@ -101,7 +101,7 @@ public class ParserTestObject extends AndroidTestCase {
 
 		Formula lookZPositionFormula = getFormulaBySensor(Sensors.OBJECT_LAYER);
 		assertEquals("Formula interpretation is not as expected (z-index)", testSprite.look.getZIndex(),
-				lookZPositionFormula.interpretInteger(testSprite));
+				lookZPositionFormula.interpretInteger(testSprite).intValue());
 
 	}
